Move root types to the root-pixie file
