+++ src/modules/account/data-store-api.js
@@ -4,7 +4,7 @@ import { mapFiles, mapFolders } from 'disklet'
 import { bridgifyObject } from 'yaob'
 
 import { type EdgeDataStore, type EdgePluginData } from '../../types/types.js'
-import { type ApiInput } from '../root.js'
+import { type ApiInput } from '../root-pixie.js'
 import {
   getStorageWalletFolder,
   hashStorageWalletFilename
