+++ src/modules/account/lobby-api.js
@@ -11,7 +11,7 @@ import {
 } from '../login/lobby.js'
 import { getStash } from '../login/login-selectors.js'
 import { sanitizeLoginStash, syncAccount } from '../login/login.js'
-import { type ApiInput } from '../root.js'
+import { type ApiInput } from '../root-pixie.js'
 import { ensureAccountExists, findAppLogin } from './account-init.js'
 
 type AppIdInfo = {
