+++ src/modules/context/context-pixie.js
@@ -4,7 +4,7 @@ import { combinePixies, stopUpdates } from 'redux-pixies'
 import { update } from 'yaob'
 
 import { type EdgeContext } from '../../types/types.js'
-import { type ApiInput } from '../root.js'
+import { type ApiInput } from '../root-pixie.js'
 import { makeContextApi } from './context-api.js'
 
 export type ContextOutput = {
