+++ src/modules/currency/wallet/currency-wallet-pixie.js
@@ -9,7 +9,7 @@ import {
   type EdgeCurrencyPlugin,
   type EdgeCurrencyWallet
 } from '../../../types/types.js'
-import { type ApiInput, type RootProps } from '../../root.js'
+import { type ApiInput, type RootProps } from '../../root-pixie.js'
 import {
   addStorageWallet,
   syncStorageWallet
