+++ src/modules/exchange/exchange-api.js
@@ -3,7 +3,7 @@
 import { bridgifyObject, onMethod } from 'yaob'
 
 import { type EdgeRateCache } from '../../types/types.js'
-import { type ApiInput } from '../root.js'
+import { type ApiInput } from '../root-pixie.js'
 import { getExchangeRate } from './exchange-selectors.js'
 
 /**
