+++ src/modules/login/create.js
@@ -5,7 +5,7 @@ import { base64 } from 'rfc4648'
 import { errorNames } from '../../types/error.js'
 import { type EdgeWalletInfo } from '../../types/types.js'
 import { encrypt } from '../../util/crypto/crypto.js'
-import { type ApiInput } from '../root.js'
+import { type ApiInput } from '../root-pixie.js'
 import { authRequest } from './authServer.js'
 import { makeKeysKit } from './keys.js'
 import { fixUsername, hashUsername } from './login-selectors.js'
