+++ src/modules/login/edge.js
@@ -9,7 +9,7 @@ import {
 } from '../../types/types.js'
 import { base58 } from '../../util/encoding.js'
 import { makeAccount } from '../account/account-init.js'
-import { type ApiInput } from '../root.js'
+import { type ApiInput } from '../root-pixie.js'
 import { makeLobby } from './lobby.js'
 import { makeLoginTree, searchTree, syncLogin } from './login.js'
 import { saveStash } from './loginStore.js'
